build-module-docs: Add tool to create "book" for module docs
This tool uses mdBook to generate a static "book" documenting all
NixOS modules.

Every module contains option documentation as well as an optional doc
header (which is taken from the first commented-out lines of each
module implementation).

This is currently a proof-of-concept, but it could turn out to be a
good idea!

build-module-docs: Refactor to avoid use of builtins.readFile