nixos/openntpd: Avoid complex import logic by duplicating a string
This particular module is the only one in the NixOS module list that
depends on another modules options to declare its own, which came up
during the evaluation of the module documentation builder as it does
not fully evaluate NixOS configuration.

The previous logic was written to avoid duplication of a single
string, which has now been inlined in the module.