stdenv: fix structuredAttrs usage on Nix 2.3
The change introducing the defaulting of the variables broke users
with Nix 2.3 who enabled structuredAttrs in stdenv builds, as the
presence of `set -u` in the file broke while checking for the presence
of the variable itself.

To work around this, the value is defaulted to an empty string if not
found and explicitly set only after that.